Fixed handleInputChange
The DateRangePicker does not return a typical event with event: {target, value}. It returns just an array.

I changed the handleInputChange function to accept two parameters, name and value (where name is the parameter to change and value is what to set it to). In all other onChanges now its a function that passes the appropriate function so the master function doesn't have to do its own weird checks :)